Rocm airbin integration (#47)
* controller: remove namespaces

Namespaces are not necessary in this project and make it more complicated to
share functions between files.

Also fix up some warnings.

Signed-off-by: Joel Nider <joel.nider@amd.com>

* controller: change hardcoded addresses to defines

Add base addresses to platform.h
Change register offsets into defines so they are easier to read.
Add some helper functions for common operations like resetting a column.

Signed-off-by: Joel Nider <joel.nider@amd.com>

* utils: script to build libxaie

Signed-off-by: Joel Nider <joel.nider@amd.com>

* platform: load & unload driver when loading fpga

Signed-off-by: Joel Nider <joel.nider@amd.com>

* Add support for libelf

Libelf (one library inside the 'elfutils' repo) is a library for manipulating
ELF files from C programs. It is used when generating AIRBIN files by the
compiler backend and also when reading them in the runtime for loading to
the device.

Signed-off-by: Joel Nider <joel.nider@amd.com>

* test: add dependency on libelf

The runtime lib was modified to use libelf for loading airbin files. Since the
runtime lib is built as a static library (.a file), it does not include a link
step and can therefore not be linked directly with libelf. However, it does
call functions in libelf when loading airbin files, which means it has a
dependency on libelf. That forces all applications that link with the runtime
lib to also link with libelf, even if those functions are not used in that
particular application.

The command line options are also put on separate lines and sorted by type
(library path, include path, etc.) to make it easier to modify, diff and
understand what is happening in general.

Signed-off-by: Joel Nider <joel.nider@amd.com>

* test: add 313 for airbin

Signed-off-by: Joel Nider <joel.nider@amd.com>

* runtime: Load airbin to device

An AIRBIN is an ELF file that contains AIE configuration and executable code.
It is loaded to the device by the runtime by copying it to device memory and
then notifying the device that continues the process.

Add code to read the file, place it in device memory and notify the device.

This depends on libelf (from elfutils) that is built in:
github-clone-build-elfutils.sh

Remove the unused 'agent' parameter from air_queue_dispatch_and_wait.

Signed-off-by: Joel Nider <joel.nider@amd.com>

* controller: add firmware support for airbin

Add a new packet type AIR_PKT_TYPE_AIRBIN for loading an AIRBIN. The packet
holds the base address in device memory where the AIRBIN has been loaded by
the runtime_lib on the host. The packet handler builds a CDMA descriptor
chain based on the AIRBIN sections and loads them into AIE memory.

The ARM_CONTROLLER define is set by the compiler so it is available to all
source files.

The CDMA driver is moved to a separate source file. Currently, only the
scatter/gather mode has been implemented. A future patch will add the simple
mode as well, and then the legacy CDMA code can be removed.

Add a function to reset a shim tile. Follows the procedure for shim tile resets.
The reset is a global action but masked by registers belonging to each shim
tile. By unmasking a particular tile, it will participate in the reset. The
actual reset is performed through the NPI interface, then all the registers are
unmasked to prevent accidental reset later.

Signed-off-by: Joel Nider <joel.nider@amd.com>

* driver: build tags

* Add one working with AIRBIN using ROCr

* Getting weather stencil working with AIRBIN on ROCr

* Firmware cleanup to get airbin working with the weather stencil

* Makefile changes to add libelf

* Removing RISCV from controller

* AIRBIN cleanup

* Updating documentation to install and point to libelf

* Updating the default firmware to be embedded in the PDI when creating the platform

Signed-off-by: Joel Nider <joel.nider@amd.com>
Co-authored-by: Joel Nider <joel.nider@amd.com>
Co-authored-by: Eddie Richter <eddier@xilinx.com>