Remove BSP dependency
The project requires several files that are supplied by Vitis. This prevents us
from modifying those files since they are copied fresh on every build. It also
ties us to the specific implementation provided by the Vitis version being used
to compile the project. Not relying on the provided BSP will also simplify and
speed up the compilation since only the necessary files are compiled.

Break the dependency on the BSP by writing a standalone makefile and
implementing the low level code such as a UART driver. Also provides printf and
the glue functions that libc requires.

Signed-off-by: Joel Nider <joel.nider@amd.com>